feat: start the session using tmux.
This is a hack to start virtiofsd before running the nix flake. It's a
requirement, and without it, `nix run` will fail to find the socket it
needs.

I had only two options, to either use `9p` types, but in that case,
everything was only accessible by root, or use `systemd` to run these
machines, which was overkill for my usecase.