refactor: switch to running the vm as a service.
Running the VMs directly, will provide a vt220 terminal, which doesn't
work well with Gemini TUI application, and causing it to flicker on
any interaction.

Running this as a service, with a preserved home directory, which
allows the host to have access to agent's home directory, proved to be
more practical. This way, users are not limited to use one machine per
project, and can mount as many project as needed to the agent.